refactor: ScheduleList 폴더 구조 변경
