--- src/components/ScheduleList/ConsumptionCard/index.ts
@@ -1,3 +1,0 @@
-import ConsumptionCard from "./ConsumptionCard.tsx";
-
-export default ConsumptionCard;
