--- src/components/ScheduleList/ConsumptionHeader/index.ts
@@ -1,3 +1,0 @@
-import ConsumptionHeader from "./ConsumptionHeader.tsx";
-
-export default ConsumptionHeader;
