+++ src/components/ScheduleList/ScheduleCard/ScheduleCard.styles.ts
