+++ src/components/ScheduleList/ScheduleCard/index.ts
@@ -1,0 +1,3 @@
+import ScheduleCard from "./ScheduleCard.tsx";
+
+export default ScheduleCard;
