+++ src/components/ScheduleList/ScheduleDateBox/index.ts
@@ -1,0 +1,3 @@
+import ScheduleDateBox from "./ScheduleDateBox.tsx";
+
+export default ScheduleDateBox;
