+++ src/components/ScheduleList/ScheduleListHeader/ScheduleListHeader.tsx
@@ -1,65 +1,37 @@
-import { Typography, Stack, Box } from "@mui/material";
-import ClearRoundedIcon from "@mui/icons-material/ClearRounded";
-import SearchIcon from "@mui/icons-material/Search";
-import TuneRoundedIcon from "@mui/icons-material/TuneRounded";
-import { useNavigate } from "react-router-dom";
-import SelectYearMonth from "@components/common/SelectYearMonth";
-import { HeaderBox } from "./ScheduleListHeader.styles";
+import { Stack, Typography } from "@mui/material";
+import SelectMenus from "@components/common/SelectMenus";
+import * as React from "react";
 
-export interface ScheduleListHeaderProps {
-  date: string;
-  title?: string;
-  addMonth: () => void;
-  subtractMonth: () => void;
-  changeMonth: () => void;
-  handleClickSearch: () => void;
-  handleClickFilter: () => void;
+export interface ReportListHeaderProps {
+  count: number;
+  options: string[];
+  selectedOption: string;
+  setSelectedOption: React.Dispatch<React.SetStateAction<string>>;
 }
 
 function ScheduleListHeader({
-  date,
-  title,
-  changeMonth,
-  subtractMonth,
-  addMonth,
-  handleClickSearch,
-  handleClickFilter,
-}: ScheduleListHeaderProps) {
-  const navigate = useNavigate();
+  count,
+  options,
+  selectedOption,
+  setSelectedOption,
+}: ReportListHeaderProps) {
   return (
-    <Box bgcolor="#735bf2">
-      <HeaderBox>
-        <Stack
-          direction="row"
-          spacing={1}
-          justifyContent="space-between"
-          alignItems="center"
-        >
-          <ClearRoundedIcon onClick={() => navigate(-1)} />
-          <Typography variant="h4" component="div" sx={{ flexGrow: 1 }}>
-            {title}
-          </Typography>
-          <SearchIcon onClick={handleClickSearch} />
-        </Stack>
-      </HeaderBox>
-
-      <HeaderBox>
-        <Stack
-          direction="row"
-          spacing={1}
-          justifyContent="space-between"
-          alignItems="center"
-        >
-          <SelectYearMonth
-            date={date}
-            lastMonth={subtractMonth}
-            nextMonth={addMonth}
-            changeYearAndMonth={changeMonth}
-          />
-          <TuneRoundedIcon onClick={handleClickFilter} />
-        </Stack>
-      </HeaderBox>
-    </Box>
+    <Stack
+      direction="row"
+      justifyContent="space-between"
+      alignItems="center"
+      py={1}
+      px={2.5}
+    >
+      <Typography>
+        총&nbsp;<span style={{ color: "#735BF2" }}>{count}</span>건
+      </Typography>
+      <SelectMenus
+        options={options}
+        selectedOption={selectedOption}
+        setSelectedOption={setSelectedOption}
+      />
+    </Stack>
   );
 }
