+++ src/components/ScheduleList/ScheduleListPageHeader/index.ts
@@ -1,0 +1,3 @@
+import ScheduleListPageHeader from "./ScheduleListPageHeader.tsx";
+
+export default ScheduleListPageHeader;
