+++ src/components/ScheduleList/index.ts
@@ -1,0 +1,3 @@
+import ScheduleList from "./ScheduleList.tsx";
+
+export default ScheduleList;
