--- src/pages/Home/next-components/ScheduleList/ScheduleListSkeleton.tsx
@@ -1,9 +1,0 @@
-import ConsumptionCardSkeleton from "@components/ScheduleList/ConsumptionCard/ConsumptionCardSkeleton.tsx";
-
-function ScheduleListSkeleton() {
-  return Array(3)
-    .fill(0)
-    .map(() => <ConsumptionCardSkeleton />);
-}
-
-export default ScheduleListSkeleton;
