--- src/pages/Home/next-components/ScheduleList/index.ts
@@ -1,3 +1,0 @@
-import ScheduleList from "./ScheduleList";
-
-export default ScheduleList;
