+++ src/pages/Home/pages/MonthSchedulePage/MonthSchedulePage.tsx
@@ -3,11 +3,11 @@ import ThickDivider from "@components/common/ThickDivider.tsx";
 import CalendarHeader from "@pages/Home/next-components/ScheduleCalendar/CalendarHeader";
 import Calendar from "pages/Home/pages/MonthSchedulePage/components/Calendar";
 import moment from "moment/moment";
-import ScheduleList from "@pages/Home/next-components/ScheduleList";
+import ScheduleList from "@components/ScheduleList";
 import useMonthSchedule from "@hooks/home/useMonthSchedule.ts";
 import MonthlyBudgetSummarySkeleton from "@pages/Home/next-components/HomeHeader/MonthlyBudgetSummary/MonthlyBudgetSummarySkeleton.tsx";
 import CalendarHeaderSkeleton from "@pages/Home/next-components/ScheduleCalendar/CalendarHeader/CalendarHeaderSkeleton.tsx";
-import ScheduleListSkeleton from "@pages/Home/next-components/ScheduleList/ScheduleListSkeleton.tsx";
+import ScheduleListSkeleton from "@components/ScheduleList/ScheduleListSkeleton.tsx";
 import { Box } from "@mui/material";
 import { HomePageProps } from "@pages/Home/Home.tsx";
 import { useEffect } from "react";
