--- src/pages/reports/ReportCategoryDetails/components/ReportListHeader/ReportListHeader.tsx
@@ -1,38 +1,0 @@
-import { Stack, Typography } from "@mui/material";
-import SelectMenus from "@components/common/SelectMenus";
-import * as React from "react";
-
-export interface ReportListHeaderProps {
-  count: number;
-  options: string[];
-  selectedOption: string;
-  setSelectedOption: React.Dispatch<React.SetStateAction<string>>;
-}
-
-function ReportListHeader({
-  count,
-  options,
-  selectedOption,
-  setSelectedOption,
-}: ReportListHeaderProps) {
-  return (
-    <Stack
-      direction="row"
-      justifyContent="space-between"
-      alignItems="center"
-      py={1}
-      px={2.5}
-    >
-      <Typography>
-        총&nbsp;<span style={{ color: "#735BF2" }}>{count}</span>건
-      </Typography>
-      <SelectMenus
-        options={options}
-        selectedOption={selectedOption}
-        setSelectedOption={setSelectedOption}
-      />
-    </Stack>
-  );
-}
-
-export default ReportListHeader;
