--- src/pages/reports/ReportCategoryDetails/components/ReportListHeader/index.ts
@@ -1,3 +1,0 @@
-import ReportListHeader from "./ReportListHeader.tsx";
-
-export default ReportListHeader;
